requirements.txt
torch
torchvision
pillow
matplotlib
scikit-learn
